+++ LSLib/Granny/Model/CurveData/D4nK8uC7u.cs
@@ -38,7 +38,7 @@ public class D4nK8uC7u : AnimationCurveData
         var numKnots = NumKnots();
         var knots = new List<float>(numKnots);
         for (var i = 0; i < numKnots; i++)
-            knots.Add((float)KnotsControls[i] / OneOverKnotScale);
+            knots.Add(KnotsControls[i] / OneOverKnotScale);
 
         return knots;
     }
