+++ LSLib/Granny/Model/CurveData/D9I1K16uC16u.cs
@@ -26,7 +26,7 @@ public class D9I1K16uC16u : AnimationCurveData
         var numKnots = NumKnots();
         var knots = new List<float>(numKnots);
         for (var i = 0; i < numKnots; i++)
-            knots.Add((float)KnotsControls[i] / scale);
+            knots.Add(KnotsControls[i] / scale);
 
         return knots;
     }
@@ -42,7 +42,7 @@ public class D9I1K16uC16u : AnimationCurveData
         var knots = new List<Matrix3>(numKnots);
         for (var i = 0; i < numKnots; i++)
         {
-            var scale = (float)KnotsControls[numKnots + i] * ControlScale + ControlOffset;
+            var scale = KnotsControls[numKnots + i] * ControlScale + ControlOffset;
             var mat = new Matrix3(
                 scale, 0, 0,
                 0, scale, 0,
